QAChat/Data_Processing/requirements.txt
langchain
nltk
deepl
python-dateutil
supabase
pytesseract
pdf2image
pdfminer.six
InstructorEmbedding
sentence-transformers
pandas
requests
slack-sdk
bs4
atlassian-python-api
spacy
google-api-python-client
pydrive2
google-auth
google-auth-oauthlib
python-dotenv
